+++ FE/src/features/schedule/create/buttons/CreateScheduleModal.tsx
